### 2022-08-23 #4701 @editorialbot
[PRE REVIEW]: enetLTS: Robust and Sparse Methods for High Dimensional Linear, Binary, and Multinomial Regression
**Submitting author:** @fatmasevinck (<a href="http://orcid.org/0000-0002-5958-7366">Fatma Sevinç Kurnaz</a>)
**Repository:** https://github.com/fatmasevinck/enetLTS
**Branch with paper.md** (empty if default branch):
**Version:** 1.1.0
