+++ paper/paper.bib
@@ -1,6 +1,7 @@
 
 
 @article{Alfons13,
+    doi     = "10.1214/12-AOAS575",
     author  = "A. Alfons and C. Croux and S. Gelper",
     title   = "Sparse least trimmed squares regression for analyzing high-dimensional large data sets",
     year    = "2013",
