### 2022-09-12 #4701 @fabian-s
:wave: @s-canchi  @mmrabe @mcavs
Would you be willing to review this submission for JOSS?
