### 2022-09-19 #4701 @fabian-s
:wave: @s-canchi @mcavs @muellsen
Would you be willing to review this submission for JOSS?
