### 2022-09-20 #4773 @editorialbot
[REVIEW]: enetLTS: Robust and Sparse Methods for High Dimensional Linear, Binary, and Multinomial Regression
**Submitting author:** @fatmasevinck (<a href="http://orcid.org/0000-0002-5958-7366">Fatma Sevinç Kurnaz</a>)
**Repository:** https://github.com/fatmasevinck/enetLTS
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
