### 2022-09-20 #4773 @fabian-s
**As a reviewer, the first step is to create a checklist for your review by entering**

```@editorialbot generate my checklist```

Please feel free to ping me (@fabian-s ) if you have any questions/concerns. Thank you for supporting JOSS with your fair and dilligent reviews!
