### 2022-10-03 #4773 @fabian-s
hi @marastadler @mcavs, thanks for filling out the review checklists.

please help @fskurnaz adress the problems you have identified by adding some more specific comments on what's missing (and how to fix them, if you have ideas), preferably by opening corresponding review issues or pull requests at https://github.com/fatmasevinck/enetLTS/issues and linking them back to this issue by including `See openjournals/joss-reviews#4773`  in the text of the issue(s)/PRs you open.

@fskurnaz:
please start adressing the problems pointed out by the reviewers as soon as possible - what is your timeframe for doing so?
