### 2022-10-04 fatmasevinck/enetLTS#14 @marastadler
Long runtime + minor comments (See openjournals/joss-reviews#4773)
Hi @fatmasevinck,

I'm in the process of reviewing `See openjournals/joss-reviews#4773`. Here are some comments on the code examples:

- The runtime of your examples is very long (for instance the multinomial example(`fit.multinom`) takes for me **18.78 mins**). You might consider improving the code to make it more time efficient. If this is not possible, you should mention in your tutorial that the runtime is quite long.

- Also, you should set a seed in your examples so that your results can be replicated.

- If you remove the ">" characters before the code lines it is more convenient for users to copy and try out your code.
