### 2022-10-04 #4773 @marastadler
Hi @fatmasevinck,
I like the general structure of your paper. The text itself still needs a little polishing.
 - In some places the quotation is in brackets, although it is part of the sentence (see, for example, lines 26/27, 42, 71, 92/93).
- Sometimes spaces are missing in the text (e.g. enetLTS()function (line 87), nCsteps (line 57)).
- Also, the manuscript should be double-checked for typos (like fuit instead of fruit, e.g. 105).
