### 2022-10-04 #4773 @marastadler
@fatmasevinck You add plots in Readme.md in the multinomial regression section that are not in the manuscript.
When plotting the coefficients in the first of these plots, the x-axis label is difficult to read. You could sort the features by importance and/or show only a subset of the coefficients.
