### 2022-10-04 #4773 @marastadler
@fatmasevinck the community guidelines are missing. One option to add such guidelines can be seen here: https://github.com/fabian-sp/GGLasso.
