### 2022-10-04 #4773 @marastadler
@fatmasevinck in the chapter "related software" you name a few packages. It would be useful to elaborate a bit more on how these packages differ from `enetLTS` (e.g. would the package `pense` in the linear model case give the same results as `enetLTS`, or are they different and, if so, why?)
