### 2022-10-05 #4773 @fabian-s
hi @mcavs, thank you for filling out the review checklist.

Please help @fatmasevinck adress the problems you have identified by adding some more specific comments on what's missing (and how to fix them, if you have ideas), preferably by opening corresponding review issues or pull requests at https://github.com/fatmasevinck/enetLTS/issues and linking them back to this issue by including See openjournals/joss-reviews#4773 in the text of the issue(s)/PRs you open.
