### 2022-10-10 #4773 @fabian-s
hi @mcavs, thank you for filling out the review checklist. JOSS aims for collaborative and constructive style reviews so I would like to ask you to please help the authors address the problems you have identified by

- adding some more specific comments on what's missing (and how to fix them, if you have ideas),
- preferably by opening corresponding review issues or pull requests at https://github.com/fatmasevinck/enetLTS/issues and
- linking them back to this issue by including See #4773 in the text of the issue(s)/PRs you open.

Thank you for supporting JOSS!
