+++ README.md
@@ -20,7 +20,7 @@
 Package `enetLTS` is on CRAN (The Comprehensive `R` Archive Network), hence the latest release can be easily installed from the `R` command as follows
 
 ```R
-> install.packages("enetLTS")
+install.packages("enetLTS")
 ```
 
  
@@ -29,8 +29,8 @@ Package `enetLTS` is on CRAN (The Comprehensive `R` Archive Network), hence the 
 The package `enetLTS` can also installed from `Github` using following the `R` command line
 
 ```R
-> install.packages("devtools")
-> devtools::install_github("fatmasevinck/enetLTS")
+install.packages("devtools")
+devtools::install_github("fatmasevinck/enetLTS")
 ```
 
 If you already have package `devtools` installed, the first line can be skipped.
@@ -45,29 +45,30 @@ Using the Affymetrix HG-U133A chip and normalizing with the GCRMA method, the nu
 As in Alfons ([Alfons, 2021](https://joss.theoj.org/papers/10.21105/joss.03786)), we determine the response variable with one of the protein expressions, which is 92th protein. Out of the gene expressions of the 22,283 genes for predictors, we have considered the gene expressions of the 100 genes that have the highest (robustly estimated) correlations with the response variable.  The code lines for loading and re-organizing response variable and predictors are follows:
 
 ```R
-> # load data
-> library("robustHD")
-> data("nci60")  # contains matrices 'protein' and 'gene'
->
-> # define response variable
-> y <- protein[, 92]
-> # screen most correlated predictor variables
-> correlations <- apply(gene, 2, corHuber, y)
-> keep <- partialOrder(abs(correlations), 100, decreasing = TRUE)
-> X <- gene[, keep]
+# load data
+library("robustHD")
+data("nci60")  # contains matrices 'protein' and 'gene'
+
+# define response variable
+y <- protein[, 92]
+# screen most correlated predictor variables
+correlations <- apply(gene, 2, corHuber, y)
+keep <- partialOrder(abs(correlations), 100, decreasing = TRUE)
+X <- gene[, keep]
 ```
 
 The package `enetLTS` can either be installed from CRAN or directly from `Github`. The main function is `enetLTS`, and the default `family` option is `gaussian`, which corresponds to linear regression.
 
 ```R
-> # install and load package
-> install.packages("enetLTS")
-> library(enetLTS)
-> # fit the model for family="gaussian"
-> fit.gaussian <- enetLTS(X,y)
-> [1] "optimal model: lambda = 0.1391 alpha = 0.6"
-> 
-> fit.gaussian
+# install and load package
+install.packages("enetLTS")
+library(enetLTS)
+# fit the model for family="gaussian"
+set.seed(4)
+fit.gaussian <- enetLTS(X,y)
+[1] "optimal model: lambda = 0.1391 alpha = 0.6"
+
+fit.gaussian
 enetLTS estimator 
 
 Call:  enetLTS(xx = X, yy = y, family = "gaussian", alphas = alphas, lambdas = lambdas, lambdaw = NULL, intercept = TRUE, scal = TRUE, 
@@ -110,17 +111,18 @@ Examples of the residuals plot (left) and the diagnostic plot (right) for output
[…]
