### 2022-10-19 fatmasevinck/enetLTS#15 @mcavs
JOSS Review for "enetLTS: Robust and Sparse Methods for High Dimensional Linear, Binary, and Multinomial Regression" #4773
Hi @fatmasevinck,

I have some comments about your paper submitted to JOSS ([openjournals/joss-reviews/issues/4773](https://github.com/openjournals/joss-reviews/issues/4773)):

* I agree with @marastadler about the long run time.
* The {enetLTS} may be cited at the beginning of the paper. I see that it is available in CRAN (you also mentioned this towards the end of the paper, but it is worth citing before) and this is a plus for it.
* There are some overflowing subtitles and code lines. Especially, some lines of the codes on page 6 are invisible. You may put the last argument to the next line.
* You may remove ">" from the code lines, or at least use the "[1]" style for code outputs to make it easier to understand for readers the codes and the outputs.

Thanks,
Mustafa
