### 2022-10-26 fatmasevinck/enetLTS#15 @fatmasevinck
Hi @mcavs

Thank you for your suggestions.
I have worked on your all suggestions, except for second one. Because the summary part starts with {enetLTS} R package information.

Best,
Sevinç
