### 2022-10-31 fatmasevinck/enetLTS#16 @fabian-s
add Contributing.md community guidelines etc
see openjournals/joss-reviews/issues/4773

See our checklist:

- [ ] Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

please reply in this issue once this has been fixed
