### 2022-10-31 #4773 @fabian-s
as far as I can determine given the lack of communication by [fatmasevinck](https://github.com/fatmasevinck), the only outstanding issue here now is

/fatmasevinck/enetLTS/issues/16 : adding a contributing.md or similar as community guidelines.

@mcavs @marastadler is this assessment correct?
