### 2022-10-31 #4773 @mcavs
@fabian-s The issues that I commented on [https://github.com/fatmasevinck/enetLTS/issues/15](https://github.com/fatmasevinck/enetLTS/issues/15) are solved by @fatmasevinck. It is ok for me.
