### 2022-10-31 fatmasevinck/enetLTS#14 @marastadler
Hi @fatmasevinck,

Thanks for addressing the comments.
After removing ">" characters from the sample code, the output of your code shouldn't be part of the code chunks. You can easily solve this by adding `eval = TRUE` to the chunk.

Please also consider addressing the comments that I've made in the review thread regarding (typos, citation, plots, community guidelines and the chapter "related software")
