### 2022-10-31 #4773 @marastadler
@fabian-s some of my comments have not been addressed yet. I have specified what is still missing in https://github.com/fatmasevinck/enetLTS/issues/14
