+++ README.md
@@ -19,7 +19,7 @@
 
 Package `enetLTS` is on CRAN (The Comprehensive `R` Archive Network), hence the latest release can be easily installed from the `R` command as follows
 
-```R
+```{R, eval = TRUE}
 install.packages("enetLTS")
 ```
 
@@ -28,7 +28,7 @@ install.packages("enetLTS")
 
 The package `enetLTS` can also installed from `Github` using following the `R` command line
 
-```R
+```{R, eval = TRUE}
 install.packages("devtools")
 devtools::install_github("fatmasevinck/enetLTS")
 ```
@@ -44,7 +44,7 @@ Using the Affymetrix HG-U133A chip and normalizing with the GCRMA method, the nu
 
 As in Alfons ([Alfons, 2021](https://joss.theoj.org/papers/10.21105/joss.03786)), we determine the response variable with one of the protein expressions, which is 92th protein. Out of the gene expressions of the 22,283 genes for predictors, we have considered the gene expressions of the 100 genes that have the highest (robustly estimated) correlations with the response variable.  The code lines for loading and re-organizing response variable and predictors are follows:
 
-```R
+```{R, eval = TRUE}
 # load data
 library("robustHD")
 data("nci60")  # contains matrices 'protein' and 'gene'
@@ -59,7 +59,7 @@ X <- gene[, keep]
 
 The package `enetLTS` can either be installed from CRAN or directly from `Github`. The main function is `enetLTS`, and the default `family` option is `gaussian`, which corresponds to linear regression.
 
-```R
+```{R, eval = TRUE}
 # install and load package
 install.packages("enetLTS")
 library(enetLTS)
@@ -104,7 +104,7 @@ Examples of the residuals plot (left) and the diagnostic plot (right) for output
 
 For binary regression, we have considered the same NCI-60 data with some modifications. In order to provide an example for binary regression, the response variable is re-organized as follows. If `mean(y)` is smaller than 0.5, the response will be assigned to 0, otherwise, the response will be assigned to 1. The predictors are the same as in the previous section.
 
-```R
+```{R, eval = TRUE}
 y <- protein[, 92]
 # for binary class 
 y.binom <- ifelse(y <= mean(y),0,1)
@@ -112,7 +112,7 @@ y.binom <- ifelse(y <= mean(y),0,1)
 
 For the binary regression, the `family` arguman of `enetLTS()`function should be set to `"binomial"`.
 
-```R
+```{R, eval = TRUE}
 # determine alpha and lambda sequences
 alphas=seq(0,1,length=41)
 l0 <- lambda00(X, y.binom, normalize = TRUE, intercept = TRUE)
@@ -147,7 +147,7 @@ Examples of plot functions of deviances (left); diagnostic (right) for binary re
 
 The fuit data set has been well-known in the context of robust discrimination studies. Therefore, we have considered the fruit data set in order to illustrate the multinomial regression. It contains spectral information with 256 wavelengths for observations from 3 different cultivars of the same fruit, named D, M, and HA, with group sizes 490, 106, and 500. This data set is available in R package `rrcov` and it is taken into consideration to illustrate the `enetLTS` model for multinomial regression.
 
-```R
+```{R, eval = TRUE}
 # load data
 library(rrcov)
 data(fruit)
@@ -161,7 +161,7 @@ y <- factor(grp-1)
 
 With `family="multinomial"`, the model `enetLTS()` produces the results of multinomial regression. Here user supplied values of `lambdas` are considered. 
 
-```R
+```{R, eval = TRUE}
 alphas=seq(from=0.01,to=0.1,by=0.01)
 set.seed(4)
 fit.multinom <- enetLTS(X, y, family="multinomial", alphas=alphas, lambdas=seq(from=0.01,to=0.1,by=0.01), crit.plot=FALSE)
