### 2022-10-31 fatmasevinck/enetLTS#14 @fatmasevinck
Dear @marastadler

I think I have done everything you have suggested, except for the last comment.
If I have missed something, your answer is very welcomed.

Thanks,
Sevinç
