### 2022-10-31 fatmasevinck/enetLTS#14 @marastadler
> Dear @marastadler
>
> Many thanks for your suggestions and comments. I am working on them but there is something I could not understand which is that "After removing ">" characters from the sample code, the output of your code shouldn't be part of the code chunks. You can easily solve this by adding eval = TRUE to the chunk." Here this R lines are only for having R format. When I added eval=TRUE to the chunk, style is broken. Maybe I missed something. If so, please could you give me more info what do you mean / is it ok in this way. Thanks again,
>
> Sevinç

I see, so you shouldn't edit the .md-file via the console here on GitHub. You have to open this file as a .Rmd-file in R and render it to a GitHub document.

This is done as follows:

title: "..."
output: github_document

Once you push your repository it will appear in the correct format.
