### 2022-11-01 fatmasevinck/enetLTS#14 @fabian-s
> Now I can see that you've fixed citations & typos in the .md file. But when downloading the article proof there are no updates. @fabian-s is it supposed to be like that?

are you referring to the `README.md` file of the repo or the JOSS paper that gets generated from `paper/paper.md` in this repo?
The PDF for the latter needs to be re-generated from the updated `paper.md` file first, you can do that in the original review thread by running `@editorialbot generate pdf` -- i did just now, this should point to the latest version of the paper: https://github.com/openjournals/joss-reviews/issues/4773#issuecomment-1298246807
