### 2022-11-01 fatmasevinck/enetLTS#14 @fatmasevinck
> > Dear @marastadler
> > I think I have done everything you have suggested, except for the last comment. If I have missed something, your answer is very welcomed.
> > Thanks, Sevinç
>
> Dear @fatmasevinck,
>
> Now I can see that you've fixed citations & typos in the .md file. But when downloading the article proof there are no updates. @fabian-s is it supposed to be like that?
>
> From what I can see my other comments on community guidelines, coefficient plots and the chapter "related software" have not been addressed yet.

Dear @marastadler

It seems you are talking about README.md file. I have written "related software" section in paper.md
Is that ok in this way?
