### 2022-11-01 fatmasevinck/enetLTS#14 @marastadler
@fatmasevinck all I'm saying is that my comment regarding the chapter "related software" has not been addressed so far.

Here's again the comment that I made:
In the chapter "related software" you name a few packages. It would be useful to elaborate a bit more on how these packages differ from enetLTS (e.g. would the package pense in the linear model case give the same results as enetLTS, or are they different and, if so, why?)
