### 2022-11-01 fatmasevinck/enetLTS#14 @fabian-s
> It would be useful to elaborate a bit more on how these packages differ from enetLTS (e.g. would the package pense in the linear model case give the same results as enetLTS, or are they different and, if so, why?)

seconded
