### 2022-11-01 fatmasevinck/enetLTS#14 @fatmasevinck
Dear @marastadler

I added some sentences to the "related software" section.

Thanks for your comments.
