### 2022-11-15 fatmasevinck/enetLTS#14 @fatmasevinck
Dear @fabian-s

Is there something that I should work on it?
I finished everything and sent my answers before.

Many thanks for your kind effort.
Sevinç
