### 2022-11-17 fatmasevinck/enetLTS#14 @fabian-s
> From what I can see my other comments on community guidelines, coefficient plots and the chapter "related software" have not been addressed yet.
