### 2022-11-17 fatmasevinck/enetLTS#14 @marastadler
@fabian-s my comments were not or only superficially addressed yet.
Here's again a summary of what's missing:
- Related software: The recently added comparison is quite superficial and also contains spelling and grammar mistakes.
- community guidelines are still missing
- I also find it still not user friendly to have both code and output (which is not commented out) in the chunks
- my comment concerning the presentation of the coefficients (third last figure in Readme.md) was not answered
