+++ CONTRIBUTING.md
@@ -1,1 +1,9 @@
-Describe here how to contribute this software?
+# Contributing to enetLTS
+In case you have questions, possible bugs or feature requests, please, open an issue that adresses your problem on Github.
+
+# Sending pull requests
+If you added some functionality, tests, or examples in the project, first create a fork in Github and make the contribution public in your repo. 
+Then send a pull request as decribed in [Creating pull requests](https://docs.github.com/en/github/collaborating-with-issues-and-pull-requests/creating-a-pull-request).
+
+# Fixing typos
+You can fix typos, spelling mistakes, or grammatical errors in the documentation directly using the GitHub web interface, as long as the changes are made in the source file.
