+++ README.md
@@ -114,7 +114,7 @@ y.binom <- ifelse(y <= mean(y),0,1)
 
 For the binary regression, the `family` arguman of `enetLTS()` function should be set to `"binomial"`.
 
-```{R, eval = TRUE}
+```{R}
 # determine alpha and lambda sequences
 alphas=seq(0,1,length=41)
 l0 <- lambda00(X, y.binom, normalize = TRUE, intercept = TRUE)
@@ -123,8 +123,11 @@ lambdas <- seq(l0,0.001,by=-0.025*l0)
 set.seed(12)
 fit.binomial <- enetLTS(X, y.binom, alphas=alphas, lambdas=lambdas, family="binomial")
 fit.binomial
-enetLTS estimator 
 
+enetLTS estimator 
+```
+
+```{R}
 Call:  enetLTS(xx = X, yy = y.binom, family = "binomial", alphas = alphas, lambdas = lambdas) 
 
  number of the nonzero coefficients:
@@ -170,9 +173,12 @@ fit.multinom <- enetLTS(X, y, family="multinomial", alphas=alphas, lambdas=seq(f
 [1] "optimal model: lambda = 0.01 alpha = 0.01"
 
 fit.mutinom 
+```
+
+```{R}
 enetLTS estimator 
 
-Call:  enetLTS(xx = xx, yy = yy, family = "multinomial", alphas = alphas, lambdas = lambdas, lambdaw = NULL, intercept = TRUE, scal = TRUE, hsize = 0.75, nsamp = c(500, 10), nCsteps = 20, nfold = 5, repl = 1, ncores = 1, tol = -1e+06, seed = NULL, crit.plot = FALSE) 
+Call:  enetLTS(xx = X, yy = y, family = "multinomial", alphas = alphas, lambdas = lambdas, seq(from=0.01,to=0.1,by=0.01), crit.plot=FALSE) 
 
  number of the nonzero coefficients:
 [1] 745
@@ -180,6 +186,45 @@ Call:  enetLTS(xx = xx, yy = yy, family = "multinomial", alphas = alphas, lambda
  alpha: 0.01
  lambda: 0.01
  lambdaw: 0.004347225
+```    
+
+
+
+```R
+# load data
+library(rrcov)
+data(fruit)
+
+d <- fruit[,-1]  # first column includes the fruit names 
+X <- as.matrix(d)
+# define response variable
+grp <- c(rep(1,490),rep(2,106),rep(3,500)) 
+y <- factor(grp-1)
+```
+With `family="multinomial"`, the model `enetLTS()` produces the results of multinomial regression. Here user supplied values of `lambdas` are considered. 
+
+```R
+lambdas=seq(from=0.01,to=0.1,by=0.01)
+set.seed(4)
+fit.multinom <- enetLTS(X, y, family="multinomial", lambdas=lambdas, 
+                        crit.plot=FALSE)
+[1] "optimal model: lambda = 0.01 alpha = 0.02"
+
+fit.mutinom 
+```
+
+```R
+enetLTS estimator 
+
+Call:  enetLTS(xx = X, yy = y, family = "multinomial", lambdas=lambdas, 
+               crit.plot = FALSE) 
+
+ number of the nonzero coefficients:
+[1] 704
+
+ alpha: 0.02
+ lambda: 0.01
+ lambdaw: 0.003971358
   ```    
 
 The main function `enetLTS()` provides similar options for the $\alpha$ sequence of the elastic net penalty. The default for the tuning parameters $\lambda$ are values from 0.95 to 0.05 with steps of size -0.05, see [Kurnaz and Filzmoser, 2022](https://arxiv.org/pdf/2205.11835.pdf).
