### 2022-12-22 e4ea9d7
Update paper.md
line 86: 
23 out of 100 independent variables are selected by the enetLTS model based on optimal combination of $\alpha=0.8$ and $\lambda=0.1043$. Here $\lambda_w =0.1824974$ corresponds to updated tuning parameter for reweighted model. 

line=142:
48 out of 100 independent variables are selected by the enetLTS model based on optimal combination of $\alpha=0.325$ and $\lambda=0.0011$. Here $\lambda_w =0.01456879$ corresponds to updated tuning parameter for reweighted model. 

line194:
704 out of 1096 independent variables are selected by the enetLTS model based on optimal combi
[…]
