### 2023-01-10 e1afbc7
Update paper.md
Code format is corrected in paper.md
