### 2023-01-12 345604b
Update README.md
Coefficient plot is deleted from README.md file considering the number of variables in this example.
