### 2023-02-02 #4773 @fabian-s
sorry but that's not how these deposits are supposed to look -- obviously, your archival deposit must be *identical* to the Git repo that was reviewed by us -- same folder structure, all the files incl. the paper,  etc, etc, and not just an unsorted, incomplete dump of program files…..

Pls do take the time to take a look at any other published JOSS paper to understand how this is supposed to look (e.g https://zenodo.org/record/7554373 for https://joss.theoj.org/papers/10.21105/joss.04848) --  you could save us all a lot of back and forth by doing a little more due dilligence or asking a more technical person from your institution to help you out if necessary.

pls remember we also need a tagged release on github so we can *finally* wrap this up.
