### 2023-02-02 #4773 @fatmasevinck
Thanks for your answer and understanding.

I have mad the tagged release yesterday (https://github.com/fatmasevinck/enetLTS/releases/tag/v0.1.0)

Now my archival deposit is identical to Git repo (https://zenodo.org/record/7598948#.Y9ur4exBxo4)

Hope that everything is ok now.
Best @fabian-s,
Sevinç
