### 2023-02-03 #4773 @fabian-s
please adjust the license of your zenodo repo (currently CC Attribution 4.0) so it matches that of your repo (GNU General Public License v3.0.)
